GET/MGET across all types in MixedProjectionFinder
Elasticsearch allows type="_all" for GET and MGET, but doesn't allow it
for SEARCH.
That's why the 'type' configured may use notations like "type1,type2"
(which may only be a subset of the available types) of the configured
'index', but '_all' will be used for GET/MGET to find a document by id
within the index. The wildcard syntax for search may also be used.