prevent non-working multi index document retrieval
as elasticsearch single index APIs such as the Document APIs do not
support multiple indices. The elasticsearch php sdk client we use
doesn't give an error on this and thus we throw an exception ourselves
when multiple indices separated by a comma or when multiple indices are
given via array.