configurable converjon thumbnail/preview activity
in the html imagelist attribute renderer to be able to change the actual
url generation for the thumbnail and preview image per attribute per view.
This is necessary when e.g. user images are stored somewhere different
from the default file storage location of all other file attributes in the
system.