add some copy/paste millisecond timings to conns
this is for local testing atm without using a debugger – when we want to
have that in the connectors we may wrap the connector calls with logging
to not repeat ourselves too often? Result mapping time etc. may be of
importance as well now that the connections are being reused via
keepalive.