BC: add getStatus() method on ConnectorInterface &
cleanup the existing implementations. The connect method of the
ConnectorInterface is now an abstract method on the abstract Connector and
doesn't exist anymore on the ConnectorService as well.

The connectors are now able to return their connection status. By default
the Status is UNKNOWN. Implementations can change the behaviour depending
on their clients/connections needs.

The RabbitMqConnector now has more configurable settings.

Most of the existing connectors now support the getStatus() method with
custom implementations. Some of them are configurable like the
elasticsearch one (ping or verbose info?) or the guzzle one (which path to
GET).