fix psr2 violations in src/
There are almost no breaking changes. I used @codingStandardsIgnoreStart
annotations on the StateMachine constructors as I didn't want to change
the order of (default) arguments there. Another place where that
annotation is used is the "_" method in the Renderer class (as that's not
camelcaps like psr2 wants it). Possible BC is the no longer present
$config=null in the Renderer's __construct method. As we use dependency
injection to create our renderers (see RendererService::createRenderer),
this should be fine (we always provide at least an empty config anyways).