remove format specific renderers and cmf specifics
The concrete output format specific renderers are used in the
honeybee-agavi-vendor-cmf repository and their twig templates are located
in that repo as well. Thus moving the renderers into the same repository
where their templates are is probably better.

A more viable solution would probably be to move everything UI related
into a standalone repository ("honeybee-ui-standard" or similar). This
could then contain renderers, twig templates and maybe even matching theme
files (scss…) and widgets (javascript behaviours…) etc.

At the moment we watch and observe how the UI related stuff grows and wait.

The position of the ServiceLocator in the honeybee lib is probably to be
discussed as well when extracting the whole Ui namespace from this lib.

refs #3