added user role based view template lookups
This enables applications to have view templates not only based on the
output format but also on the role id of a user.

A UserInterface has been introduced to make sure a getRoleId() method
exists on the Environment->getUser(). The UserInterface::isAllowed()
method was added so e.g. renderer can check whether the user with the
role is allowed to see or change the rendered subject.