add _id to exception to ease error interpretation
it might happen that non-projection documents are added to an index. When
this happens seemingly random a query for stuff throws an exception as the
document doesn't have the @type set and thus has an internal elasticsearch
_id. Showing that id should point devs to the fact that docs exist in the
index that don't adhere to the _id format and the necessary data format.