fix multi field convention for suggestions
for the search criteria having a syntax of "suggest:FIELD=VALUE" we need
to expand the FIELD name to "FIELD.suggest" as it is our convention to
have an elasticsearch  multi field with that suffix for suggestions.

This is a possible breaking change for devs using the suggestions without
having a ".suggest" multi field in their projects.

fixes #77