allow overriding CLI user for application
You may use APP_USER to set the username used for the console auth
provider. The APP_USER name is the default and may be overridden via
ConsoleAuthService configuration 'user_env_var'.

In the vendor-agavi-cmf you're allow to do e.g.:

APP_USER=my-other-user bin/cli some.route -param 'bar'

when the "my-other-user" is defined in the "role_map" of the service.