add "__notempty" as another flag for queries
It's using the wildcard and missing filters to query for documents that
have a non-empty string as a field value. The already existing "__empty"
flag in combination with "!" ("!__empty") is not the same as empty
strings are non-null values for elasticsearch and thus the field is not
empty when the string is of zero length. Beware that the wildcard filter
might not be the most efficient one for large indexes.