WIP: setting "index_per_type" for elasticsearch7+
support (on the connector/client). When that setting is present on the
connector and is "true", then the elasticsearch related classes for
migrations, storage and finders assume one type per index. This is
different from the multiple types per index approach honeybee projects
probably use until now. This change should be non-breaking. When
switching the setting on, projects must make sure, that they don't
specify multiple types in any elasticsearch migration as only the first
one of the specified typemapping paths are taken into account.